requirements.txt
bintrees==2.0.7
pygtrie==2.3.2
python-chess==0.28.3
splay-mod==1.12
tqdm==4.40.2
